Polishing the sign in and menus.
I created a RootViewController class to manage signing in and out and the
reveal controller and its menu. If we have stored credentials we skip the sign
in form and go directly to the content. There's now a nice flip transition
between the content and the sign in view. The menu also keeps the currently
selected view highlighted.

Behind the scenes I'm using a lot more notificationsto handle sign in, sign out
and household changes in a sane manner. For instance now the list views won't
fetch data if there's no session they'll wait until the RMHouseholdSelected
notice to arrives after the sign in.