Reworked the Add Note screen.
It's now in a table with an "Add Photo" row like the Add Expenses. Photos are
now stored in the data objects. It cleaned up the posting functions for both
notes and expenses. I'm trying to give them both a consistent interface so it's
easy to call them after taking a photo.